+++ src/app/core/services/constant.service.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from '@angular/core';
+
+export const ConstantService = new InjectionToken<any>('constants value');