+++ src/app/core/services/index.ts
@@ -1,2 +1,4 @@
 export * from './auth.service';
 export * from './local-storage.service';
+export * from './config-option.service.ts.service';
+export * from './constant.service';