+++ changelog.md
@@ -89,3 +89,6 @@
 # 5 (missed tasks from 3 tasks)
 
 4.1 Создан сервис ConfigOptionsService который хранит объект настроек (id, login, email...), у сервиса есть метод для установки данных который на вход принимает объект и метод извлечения данных. Предусмотрена возможность установки подмножества свойств.
+
+4.2 Создан сервис ConstantService (core/service/constant.service.ts) в виде готового литерала объекта, сервис зарегистрирован в CoreModel используя useValue: { App: 'ngShop', Ver: '1.0'} 
+4.7 В компонент AboutComponent внедрены сервисы ConfigOptionService, ConstantService, GeneratorService, GeneratorFactoryService используя декоратор @Option